+++ R/cohens_d.R
@@ -5,24 +5,23 @@
 #' For sample sizes > 20, the results for both statistics are roughly equivalent.
 #' \cr\cr
 #'  The Glass’s delta is appropriate if standard deviations are significantly different
-#'  between groups, as it uses only the \emph{second} group's standard deviation.
+#'  between groups, as it uses only the *second* group's standard deviation.
 #'
-#' @param x A formula, a numeric vector, or a name of one in \code{data}.
-#' @param y A numeric vector, a grouping (character / factor) vector, a or a name of one in \code{data}. Ignored if \code{x} is a formula.
+#' @param x A formula, a numeric vector, or a character name of one in `data`.
+#' @param y A numeric vector, a grouping (character / factor) vector, a or a character  name of one in `data`. Ignored if `x` is a formula.
 #' @param data An optional data frame containing the variables.
-#' @param correction If \code{TRUE}, applies a correction to make it less biased for small samples (McGrath & Meyer, 2006).
-#' @param pooled_sd If \code{TRUE} (default), a \code{\link{sd_pooled}} is used (assuming equal variance). Else the mean SD from both groups is used instead.
-#' @param paired If \code{TRUE}, the values of \code{x} and \code{y} are considered as paired.
+#' @param correction If `TRUE`, applies a correction to make it less biased for small samples (McGrath & Meyer, 2006).
+#' @param pooled_sd If `TRUE` (default), a [sd_pooled()] is used (assuming equal variance). Else the mean SD from both groups is used instead.
+#' @param paired If `TRUE`, the values of `x` and `y` are considered as paired.
 #' @inheritParams chisq_to_phi
 #'
 #' @return A data frame with the effect size(s) and confidence interval(s).
 #'
-#' \subsection{Confidence Intervals}{
+#' ## Confidence Intervals
 #' Confidence intervals are estimated using the Noncentrality parameter method;
-#' These methods searches for a the best \code{ncp} (non-central parameters) for
+#' These methods searches for a the best `ncp` (non-central parameters) for
 #' of the noncentral t distribution for the desired tail-probabilities,
-#' and then convert these \code{ncp}s to the corresponding effect sizes.
-#' }
+#' and then convert these `ncp`s to the corresponding effect sizes.
 #'
 #' @examples
 #' cohens_d(iris$Sepal.Length, iris$Sepal.Width)
@@ -33,11 +32,11 @@
 #' hedges_g(mpg ~ am, data = mtcars)
 #' glass_delta(mpg ~ am, data = mtcars)
 #'
-#' @references \itemize{
-#'  \item Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.
-#'  \item McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
-#'  \item Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.
-#' }
+#' @references
+#' - Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.
+#' - McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
+#' - Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.
+#'
 #' @importFrom stats var model.frame
 #' @export
 cohens_d <- function(x,
@@ -95,7 +94,7 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
 
 
 
-#' @importFrom stats sd
+#' @importFrom stats sd na.omit
 #' @keywords internal
 .effect_size_difference <- function(x,
                                     y = NULL,
@@ -121,14 +120,14 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
   if (paired) {
     d <- mean(x - y, na.rm = TRUE)
     s <- stats::sd(x - y, na.rm = TRUE)
-    n <- length(x)
+    n <- length(stats::na.omit(x - y))
     df <- n - 1
     hn <- 1 / df
     t <- d / (s / sqrt(n))
   } else {
     d <- mean(x, na.rm = TRUE) - mean(y, na.rm = TRUE)
-    n1 <- length(x)
-    n2 <- length(y)
+    n1 <- length(stats::na.omit(x))
+    n2 <- length(stats::na.omit(y))
     n <- n1 + n2
     df <- n - 2
     hn <- (1 / n1 + 1 / n2)
@@ -140,7 +139,6 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
         s1 <- stats::sd(x, na.rm = TRUE)
         s2 <- stats::sd(y, na.rm = TRUE)
         s <- sqrt((s1 ^ 2 + s2 ^ 2) / 2)
-        # s <- stats::sd(c(x, y), na.rm = TRUE)
         t <- d / sqrt(s1 ^ 2 / n1 + s2 ^ 2 / n2)
       }
     } else if (type == "delta") {